feat(back) Arquivos do back e API atualizados

+++ backend/src/main/java/com/borasio_back/backend/repository/PagamentoRepository.java
@@ -3,5 +3,5 @@ package com.borasio_back.backend.repository;
 import com.borasio_back.backend.model.entity.Pagamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagamentoRepository extends JpaRepository<Pagamento, Integer> {
+public interface PagamentoRepository extends JpaRepository<Pagamento, Long> {
 }
